fix: change Standard plan platform fee from 6% to 1.4%
- Update EdBrio platform fee: Standard 6% -> 1.4% (total with Stripe: 5.0%)
- Recalculate breakeven point: ~¥26,500/month (was ¥148,000)
- Update comparison table with new fee examples
- Fix parseInt -> parseFloat in calculateApplicationFee for decimal rates
- Update LP, profile page, and legal page fee references

+++ src/app/api/checkout/session/route.ts
@@ -65,8 +65,8 @@ export async function POST(req: NextRequest) {
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 function calculateApplicationFee(amount: number, teacherPlan: 'free' | 'pro' = 'free'): number {
   const feePercent = teacherPlan === 'pro' ?
-    parseInt(process.env.EDBRIO_PLATFORM_FEE_PERCENT_PRO!) :
-    parseInt(process.env.EDBRIO_PLATFORM_FEE_PERCENT_FREE!)
+    parseFloat(process.env.EDBRIO_PLATFORM_FEE_PERCENT_PRO!) :
+    parseFloat(process.env.EDBRIO_PLATFORM_FEE_PERCENT_FREE!)
 
   const minFee = parseInt(process.env.EDBRIO_MIN_FEE_JPY!) * 100 // Convert to cents
   const calculatedFee = Math.round(amount * (feePercent / 100))
